Generates Github URLs

Generates a markdown section on where to ask questions about a package.

makeGithubUrl(pkg, user = options()[["GitHubUserName"]])

makeGithubPageUrl(pkg, user = options()[["GitHubUserName"]])

makeGithubIssuesUrl(pkg, user = options()[["GitHubUserName"]])

pkgqsts(pkg, github = TRUE, bioc = TRUE, level = 1L,
  user = options()[["GitHubUserName"]])

Arguments

pkg

Mandatory character with the package name. Must be of length one for codepkgqsts. Other functions makeGithubPageUrl and makeGithubIssuesUrl are vectorised.

user

Mandatory GitHub username. Default is to retrieve options()[["GitHubUserName"]].

github

A logical (default is TRUE) defining whether users can ask questions on GitHub issues.

bioc

A logical (default is TRUE) defining whether users can ask questions on the Bioconductor support forum.

level

Level for markdown sections. Default is 1L for #. Use 2L for ## and so on.

Value

A character with Github URLs.

A character with Github page URLs.

A character with Github page URLs.

Invisibly returns NULL. Used for its side effect.

See also

BiocStyle::Githubpkg

Examples

makeGithubUrl("pRoloc", user = "lgatto")

#>                             pRoloc 
#> "https://github.com/lgatto/pRoloc" 

makeGithubUrl(c("pRoloc", "MSnbase"), user = c("lgatto", "lgatto"))

#>                              pRoloc                             MSnbase 
#>  "https://github.com/lgatto/pRoloc" "https://github.com/lgatto/MSnbase" 

makeGithubPageUrl("pRoloc", user = "lgatto")

#>                            pRoloc 
#> "https://lgatto.github.io/pRoloc" 

makeGithubPageUrl(c("pRoloc", "pRolocGUI"), user = c("lgatto", "lgatto"))

#>                               pRoloc                            pRolocGUI 
#>    "https://lgatto.github.io/pRoloc" "https://lgatto.github.io/pRolocGUI" 

makeGithubIssuesUrl("pRoloc", user = "lgatto")

#>                                    pRoloc 
#> "https://github.com/lgatto/pRoloc/issues" 

pkgqsts("pRoloc", user = "lgatto")

#> # Questions and support
#> To get help:
#>  - Open a GitHub [issue](https://github.com/lgatto/pRoloc/issues)
#>  - Post your question on the [Bioconductor support site](https://support.bioconductor.org/)

pkgqsts("cputools", bioc = FALSE, user = "lgatto")

#> # Questions and support
#> To get help:
#>  - Open a GitHub [issue](https://github.com/lgatto/cputools/issues)

pkgqsts()

#> # Questions and support
#> To get help:
#>  - Post your question on the [Bioconductor support site](https://support.bioconductor.org/)