Generate the markdown code the a package shields.

Functions to generate package travis-ci, codecov, Bioconductor build status and Biocondcutor test coverage shields

makeTravisShield(pkg, user = options()[["GitHubUserName"]],
  branch = "master")

makeCodecovShield(pkg, user = options()[["GitHubUserName"]],
  branch = "master")

makeBiocBuildShield(pkg, branch = c("devel", "release"))

makeBiocCovrShield(pkg, branch = c("devel", "release"))

Arguments

pkg

Mandatory character with the package name(s). If of length greater than one, then must of of same length as user and branch.

user

Mandatory GitHub username. Default is to retrieve options()[["GitHubUserName"]]. Argument is not recycled of more than one package is provided.

branch

Mandatory branch name. Default is "master". Argument is not recycled of more than one package is provided. For Bioconductor shields (functions makeBiocBuildShield and makeBiocCovrShield), branch is one of "devel" (default) or "release".

Value

Invisibly returns character with the shield in markdown. Used for its side effect of printing the markdown shield.

Examples

makeTravisShield("pRoloc", user = "lgatto")

#> [![Build Status](https://travis-ci.org/lgatto/pRoloc.svg?branch=master)](https://travis-ci.org/lgatto/pRoloc) 

## default user
options(GitHubUserName = "ComputationalProteomicsUnit")
makeCodecovShield("cputools")

#> [![codecov](https://codecov.io/gh/ComputationalProteomicsUnit/cputools/branch/master/graph/badge.svg)](https://codecov.io/gh/ComputationalProteomicsUnit/cputools) 

makeCodecovShield("MSnbase", user = "lgatto")

#> [![codecov](https://codecov.io/gh/lgatto/MSnbase/branch/master/graph/badge.svg)](https://codecov.io/gh/lgatto/MSnbase) 

makeBiocBuildShield("MSnbase")

#> [![Bioconductor devel build status](http://bioconductor.org/shields/build/devel/bioc/MSnbase.svg)](http://bioconductor.org/packages/devel/bioc/html/MSnbase.html) 

makeBiocBuildShield("MSnbase", branch = "release")

#> [![Bioconductor release build status](http://bioconductor.org/shields/build/release/bioc/MSnbase.svg)](http://bioconductor.org/packages/release/bioc/html/MSnbase.html) 

makeBiocCovrShield("MSnbase")

#> [![Bioconductor devel build status](https://bioconductor.org/shields/coverage/devel/MSnbase.svg)](https://codecov.io/github/Bioconductor-mirror/MSnbase/branch/master) 

makeBiocCovrShield("MSnbase", branch = "release")

#> [![Bioconductor release build status](https://bioconductor.org/shields/coverage/release/MSnbase.svg)](https://codecov.io/github/Bioconductor-mirror/MSnbase/branch/release-3.4)